# cmd/api/docs/swagger.yaml
basePath: /
definitions:
  models.AccountEditReq:
    properties:
      color:
        example: color
        maxLength: 10
        type: string
      name:
        example: name
        maxLength: 100
        type: string
    required:
    - color
    - name
    type: object
  models.AccountGetListReq:
    properties:
      isManinAccount:
        example: true
        type: boolean
      page:
        example: 1
        minimum: 1
        type: integer
      perPage:
        example: 10
        minimum: 1
        type: integer
      searchText:
        example: search by type or name
        type: string
      sortBy:
        properties:
          field:
            example: updatedDate
            type: string
          mode:
            example: desc
            type: string
        type: object
    type: object
  models.AuthLoginReq:
    properties:
      password:
        example: "123456"
        maxLength: 6
        minLength: 6
        type: string
      username:
        example: admin
        maxLength: 100
        type: string
    required:
    - password
    - username
    type: object
  models.AuthRefreshReq:
    properties:
      refToken:
        type: string
    type: object
  models.AuthRegisterReq:
    properties:
      password:
        example: "123456"
        maxLength: 6
        minLength: 6
        type: string
      username:
        example: admin
        maxLength: 100
        type: string
    required:
    - password
    - username
    type: object
  models.BannerGetListReq:
    properties:
      page:
        example: 1
        minimum: 1
        type: integer
      perPage:
        example: 10
        minimum: 1
        type: integer
      searchText:
        example: search by title
        type: string
      sortBy:
        properties:
          field:
            example: updatedDate
            type: string
          mode:
            example: desc
            type: string
        type: object
    type: object
  models.DebitCardGetListReq:
    properties:
      page:
        example: 1
        minimum: 1
        type: integer
      perPage:
        example: 10
        minimum: 1
        type: integer
      searchText:
        example: search by name
        type: string
      sortBy:
        properties:
          field:
            example: updatedDate
            type: string
          mode:
            example: desc
            type: string
        type: object
      status:
        example: ""
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: This is a sample server for a github.com/kongzyeons/go-bank
  title: API-BANK
  version: "1.0"
paths:
  /api/v1/account/edit/{accountID}:
    put:
      consumes:
      - application/json
      description: Edit
      operationId: AccountEdit
      parameters:
      - description: accountID
        in: path
        name: accountID
        required: true
        type: string
      - description: request body
        in: body
        name: AccountEditReq
        required: true
        schema:
          $ref: '#/definitions/models.AccountEditReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Edit
      tags:
      - Account API
  /api/v1/account/getlist:
    post:
      consumes:
      - application/json
      description: GetList
      operationId: AccountGetList
      parameters:
      - description: request body
        in: body
        name: AccountGetListReq
        required: true
        schema:
          $ref: '#/definitions/models.AccountGetListReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetList
      tags:
      - HomePage API
  /api/v1/auth/logout:
    post:
      consumes:
      - application/json
      description: Logout
      operationId: AuthLogout
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Logout
      tags:
      - Auth API
  /api/v1/auth/ping:
    get:
      consumes:
      - application/json
      description: Ping
      operationId: AuthPing
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Ping
      tags:
      - Auth API
  /api/v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh
      operationId: AuthRefresh
      parameters:
      - description: request body
        in: body
        name: AuthRefreshReq
        required: true
        schema:
          $ref: '#/definitions/models.AuthRefreshReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Refresh
      tags:
      - Auth API
  /api/v1/banner/getlist:
    post:
      consumes:
      - application/json
      description: GetList
      operationId: BannerGetList
      parameters:
      - description: request body
        in: body
        name: BannerGetListReq
        required: true
        schema:
          $ref: '#/definitions/models.BannerGetListReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetList
      tags:
      - HomePage API
  /api/v1/debitCard/getlist:
    post:
      consumes:
      - application/json
      description: GetList
      operationId: debitCardGetList
      parameters:
      - description: request body
        in: body
        name: DebitCardGetListReq
        required: true
        schema:
          $ref: '#/definitions/models.DebitCardGetListReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetList
      tags:
      - HomePage API
  /api/v1/login:
    post:
      consumes:
      - application/json
      description: Login
      operationId: AuthLogin
      parameters:
      - description: request body
        in: body
        name: AuthLoginReq
        required: true
        schema:
          $ref: '#/definitions/models.AuthLoginReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Login
      tags:
      - Auth API
  /api/v1/register:
    post:
      consumes:
      - application/json
      description: Register
      operationId: AuthRegister
      parameters:
      - description: request body
        in: body
        name: AuthRegisterReq
        required: true
        schema:
          $ref: '#/definitions/models.AuthRegisterReq'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Register
      tags:
      - Auth API
  /api/v1/user/greeting:
    get:
      consumes:
      - application/json
      description: GetGeeting
      operationId: UserGetGeeting
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: GetGeeting
      tags:
      - HomePage API
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"